refactor: move salary joi schema to common file
